feat(validate): extract spec validation into a public 'validate' package
The kin-openapi error mapping logic (rule IDs, severities, source
locations, fingerprints) has lived in internal/validate.go since
PR #894 introduced 'oasdiff validate'. That keeps it locked inside
the CLI; any other consumer (oasdiff-service, third-party tools) has
no way to call it.

Move the pure validation+mapping logic into a new 'validate' package
with one public entry point:

  func Validate(spec *openapi3.T, source string) formatters.Findings

This mirrors how diff.Get and checker.CheckBackwardCompatibility are
callable from any consumer rather than being CLI-internal.

What moves:
- mapKinErrors, flattenKinErrors, dedupePreferringComponents
- severityForKinError, sectionForKinError, sectionFromField
- pathOperationForKinError, unwrapContext
- argsForKinError, lineForKinError, columnForKinError
- locationForKinError, fieldLoc
- ruleIDForKinError, joinFieldsForRuleID, ruleIDFromField
- unknownValidationID const

What stays in internal/validate.go (CLI-only):
- getValidateCmd (cobra command + flags)
- runValidate (load spec, call validate.Validate, format output)
- outputFindings (formatter lookup + render)

Behaviour unchanged: the CLI command now calls validate.Validate(spec,
source) and threads the result through the same formatter logic. All
existing CLI validate tests (internal/validate_test.go, 471 lines)
pass without modification.

While here: renamed the local validate() helper in internal/viper.go
to validateViperConfig() since 'validate' was a generic name that
collided with the new package import. One call site updated.

Blocks: oasdiff-service#200, which will call validate.Validate from
the new /public/validate endpoint.
